# api/src/routes/employment.js
const express = require('express');
const { requireScope, requireRole } = require('../middleware/auth');
const { MASTER_DATA_ADMIN_ROLE } = require('../constants');
const { requireInactiveScope } = require('../middleware/inactiveScope');
const { getEmploymentHistory, upsertEmployment } = require('../services/employmentService');

function createEmploymentRouter(pool) {
  const router = express.Router();

  router.get('/persons/:personId/employment', requireScope('personnel:read:employment'), requireInactiveScope(pool), async (req, res, next) => {
    try {
      const currentOnly = req.query.currentOnly === 'true' || req.query.currentOnly === true;
      const history = await getEmploymentHistory(pool, req.params.personId, currentOnly);
      res.json(history);
    } catch (err) {
      next(err);
    }
  });

  router.put('/persons/:personId/employment', requireScope('personnel:write:employment'), requireRole(MASTER_DATA_ADMIN_ROLE), async (req, res, next) => {
    try {
      const result = await upsertEmployment(pool, req.params.personId, req.body, req.auth);
      res.json(result);
    } catch (err) {
      next(err);
    }
  });

  return router;
}

module.exports = createEmploymentRouter;