# docs/design/personnel-mdm-openapi.yaml
openapi: 3.1.0
info:
  title: Lampang PAO - Central Personnel Master Data API
  version: 1.0.0
  summary: Single Source of Truth for personnel data of Lampang Provincial Administrative Organization (อบจ.ลำปาง)
  description: |
    API กลางสำหรับข้อมูลข้าราชการและพนักงาน อบจ.ลำปาง (Central Personnel MDM)

    **หลักการที่ผู้เรียกใช้ต้องทราบ**
    - ระบบปลายทางอ้างอิงบุคลากรด้วย `personId` (UUID) เท่านั้น เลขบัตรประชาชน (pid) จะไม่ปรากฏใน response ใด ๆ
      ยกเว้น `POST /persons/{personId}/pid` ซึ่งต้องมี scope `personnel:read:pid` และต้องระบุเหตุผลทุกครั้ง
    - การเข้าถึงข้อมูลควบคุมแบบ field-level ด้วย OAuth2 scope จาก Keycloak
      แต่ละกลุ่มฟิลด์ระบุ scope ที่ต้องใช้ไว้ใน `x-required-scope` ของ schema ฟิลด์ที่ไม่มีสิทธิ์จะถูกตัดออกจาก response (ไม่ใช่ null)
    - บุคคลสถานะ `INACTIVE` (ลาออก/โอนย้าย) เห็นได้เฉพาะ token ที่มี scope `personnel:read:inactive` ทั้งใน `GET /persons` และ endpoint รายบุคคล
      (`GET /persons/{id}`, `/photo`, `/employment`, `/change-log`, `/manage-profile`, `/history`, `POST /persons/{id}/pid`) ไม่มี scope -> 403 `insufficient-scope`
      (ไม่เขียน access log) บุคคลที่ไม่มีอยู่ยังได้ 404
    - ทุกคำขอที่คืนข้อมูลส่วนบุคคลถูกบันทึกใน access log (PDPA audit trail) พร้อม `purpose` ของระบบผู้เรียก
    - ข้อมูลระบุตัวตน (ชื่อ-สกุล วันเกิด ที่อยู่ตามทะเบียนบ้าน รูปถ่าย) มาจาก ThaID เท่านั้น แก้ไขผ่าน API ไม่ได้
    - เหตุการณ์เปลี่ยนแปลงข้อมูลส่งถึงระบบปลายทางสองทาง: webhook (push, มีลายเซ็น HMAC) และ `GET /events` (pull ด้วย cursor)
...

...
  contact:
...
servers:
...
security:
  - keycloak: []
...
paths:
...
  /persons/{personId}/employment:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Employment]
      summary: Get employment history
      operationId: getEmployment
      security:
        - keycloak: [personnel:read:employment]
      parameters:
        - name: currentOnly
          in: query
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/PidFormat'
      responses:
        '200':
          description: รายการ employment (ปัจจุบันและประวัติ)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Employment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Employment]
      summary: Upsert current employment (HR)
      description: |
        HR บันทึกหรือแก้ไขข้อมูลการปฏิบัติงานปัจจุบัน หากมีการเปลี่ยนตำแหน่ง/สังกัด ระบบจะปิด record เดิม (`effectiveTo`)
        และเปิด record ใหม่โดยอัตโนมัติ (เก็บประวัติ) และส่งเหตุการณ์ `EMPLOYMENT_UPDATED` (ไม่ส่งสำหรับบุคคลที่ยัง `PENDING_CLAIM`;
        payload ใช้สถานะจริงของบุคคล)

        **สิทธิ์ (เปลี่ยนใน PR-D1):** ต้องมี scope ของ operation นี้ **และ** realm role `hr_master_data_admin` (ตรวจที่ MDM API ทุก request
        เพราะ scope ใน token ไม่ใช่ตัวกั้นสิทธิ์) ไม่มี role -> 403 `insufficient-role` และต้องมี `reason` (บังคับ ห้ามมีเลข 13 หลัก:
        ว่าง -> 422 `reason-required`, มีเลข 13 หลัก -> 422 `reason-contains-pid`, เกิน 500 ตัวอักษร -> 422 `reason-too-long`) ซึ่งถูกเก็บใน
        `data_change_log.reason`
      operationId: upsertEmployment
      security:
        - keycloak: [personnel:write:employment]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmploymentChangeRequest' }
      responses:
        '200':
          description: บันทึกแล้ว
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Employment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
  /persons/{personId}/change-log:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Audit]
      summary: Field-level change history
      description: ประวัติการเปลี่ยนแปลงรายฟิลด์ (ฟิลด์ใด เปลี่ยนจากอะไรเป็นอะไร เมื่อไหร่ โดยใคร/แหล่งใด) ค่าของฟิลด์ที่จัดชั้น RESTRICTED จะถูก mask
      operationId: getPersonChangeLog
      security:
        - keycloak: [audit:read]
      parameters:
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: รายการการเปลี่ยนแปลง
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ChangeLogEntry' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ------------------------------------------------------- Manage (HR, PR-D2)
...
  /audit/change-logs:
    get:
      tags: [Audit]
      summary: Change history across persons / master data (DPO / auditor)
      description: |
        ประวัติการเปลี่ยนแปลงข้อมูลทั้งระบบ เรียงใหม่ -> เก่า เลือกตารางด้วย `source`:
        `PERSON` (audit.data_change_log - ข้อมูลบุคคล, กรองด้วย `personId`/`changedBy`) หรือ
        `REFERENCE` (audit.reference_change_log - หน่วยงาน/ตำแหน่ง, กรองด้วย `action`)
        ฟิลด์ชั้น CONFIDENTIAL/SENSITIVE/RESTRICTED ปกปิดค่า (`oldValue`/`newValue` = null, `valuesHidden` = true)
        เหลือเฉพาะชื่อฟิลด์ที่เปลี่ยน; ข้อความทุกช่องถูกปกปิดเลข 13 หลัก
        แถวที่เขียนก่อน PR-A ไม่มี `actorSub`/`actorClient` (null)
      operationId: listChangeLogs
      security:
        - keycloak: [audit:read]
      parameters:
        - name: source
          in: query
          schema: { type: string, enum: [PERSON, REFERENCE], default: PERSON }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: actorSub
          in: query
          schema: { type: string }
        - name: tableName
          in: query
          description: เช่น employment, person, person_contact, person_identity (PERSON) หรือ org_unit, position (REFERENCE)
          schema: { type: string }
        - name: personId
          in: query
          description: เฉพาะ source=PERSON
          schema: { type: string, format: uuid }
        - name: changedBy
          in: query
          description: เฉพาะ source=PERSON
          schema: { type: string, enum: [THAID_SYNC, SELF, HR, HR_IMPORT, ADMIN] }
        - name: action
          in: query
          description: เฉพาะ source=REFERENCE
          schema: { type: string, enum: [CREATE, UPDATE] }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: รายการ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AuditChangeLogEntry' }
                  page: { $ref: '#/components/schemas/PageInfo' }
  /audit/access-logs:
    get:
      tags: [Audit]
      summary: Who accessed whose data (DPO / HR)
      operationId: listAccessLogs
      security:
        - keycloak: [audit:read]
      parameters:
        - name: personId
          in: query
          schema: { type: string, format: uuid }
        - name: clientId
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: pidAccessOnly
          in: query
          description: เฉพาะการเข้าถึง pid / lookup
          schema: { type: boolean, default: false }
        - name: reviewStatus
          in: query
          description: |
            เฉพาะรายการ "เปิดเลขบัตร" (`POST /persons/{id}/pid` ที่สำเร็จ และแถวเก่าที่เป็น `GET`) ที่สถานะรีวิวล่าสุดตรงค่านี้
            `PENDING` = ยังไม่มีผลรีวิว (ไม่เก็บเป็นแถว คำนวณตอนอ่าน) เรียงเก่า -> ใหม่
          schema: { type: string, enum: [PENDING, REVIEWED, NEEDS_EXPLANATION] }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: รายการ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AccessLogEntry' }
                  page: { $ref: '#/components/schemas/PageInfo' }
  /audit/access-logs/{accessId}/review:
    post:
      tags: [Audit]
      summary: Record a review of a national ID reveal (DPO only)
      description: |
        บันทึกผลรีวิวรายการเปิดเลขบัตร (append-only: รีวิวซ้ำได้ ผลล่าสุดคือสถานะปัจจุบัน)
        - ต้องมี scope `audit:review` **และ** realm role `dpo` (auditor อ่านอย่างเดียว)
        - รีวิวรายการที่ตนเองเป็นผู้เปิดไม่ได้ (403 `self-review-forbidden`)
        - รายการต้องเป็นการเปิดเลขบัตรที่สำเร็จ (ไม่ใช่รายการ access log อื่น) มิฉะนั้น 404
        - `accessedAt` ใช้คู่กับ `accessId` ระบุแถว (PK ของ access_log เป็น composite เพราะ partition) ความละเอียดระดับ ms
        - `NEEDS_EXPLANATION` ต้องมี `note`; `note` ห้ามมีเลข 13 หลัก (422)
      operationId: reviewPidAccess
      security:
        - keycloak: [audit:review]
      parameters:
        - name: accessId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessedAt, status]
              properties:
                accessedAt: { type: string, format: date-time }
                status: { type: string, enum: [REVIEWED, NEEDS_EXPLANATION] }
                note: { type: string, maxLength: 1000 }
      responses:
        '201':
          description: บันทึกผลรีวิวแล้ว
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PidAccessReview' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: ไม่ได้ระบุ note ที่จำเป็น หรือ note มีเลข 13 หลัก
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /audit/alerts:
    get:
      tags: [Audit]
      summary: Access anomaly alerts (DPO / auditor)
      description: |
        แจ้งเตือนพฤติกรรมการเข้าถึงข้อมูลผิดปกติที่ worker job `access-anomaly-scan` ตรวจพบ (เรียงใหม่ -> เก่า)
        กฎ: `BULK_VIEW` (เปิดดูบุคคลหลายคนในเวลาสั้น), `OFF_HOURS` (นอกเวลาราชการ จ.-ศ. 08:30-16:30 เวลาไทย รวมเสาร์-อาทิตย์),
        `PID_REVEAL_FREQUENT` (เปิดเลขบัตรบ่อย) - ค่าเกณฑ์ตั้งจาก env ของ worker
        สถานะคำนวณจากการกระทำล่าสุด: ไม่มี = `OPEN`, รับทราบ = `ACK`, ปิดเรื่อง = `CLOSED`
        `details` มีเฉพาะตัวเลข/พารามิเตอร์ของกฎ ไม่มี pid ไม่มีชื่อ และไม่มี personId ของผู้ถูกเข้าถึง (ไล่ดูได้จาก access log ด้วย actor + ช่วงเวลา)
      operationId: listAccessAlerts
      security:
        - keycloak: [audit:read]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, ACK, CLOSED] }
        - name: ruleCode
          in: query
          schema: { type: string, enum: [BULK_VIEW, OFF_HOURS, PID_REVEAL_FREQUENT] }
        - name: from
          in: query
          description: ตรวจพบตั้งแต่ (detectedAt)
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: actorSub
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: รายการ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AccessAlert' }
                  page: { $ref: '#/components/schemas/PageInfo' }
...
components:
  securitySchemes:
    keycloak:
      type: oauth2
      description: |
        Keycloak realm `lp-pao` เป็นผู้ออก token สำหรับเรียก MDM API (RS256, ตรวจสอบด้วย JWKS) โดยแยกจากเส้นทางล็อกอินของผู้ใช้
        ซึ่งยังคงเป็น thaid.lp-pao.go.th → check.lp-pao.go.th ตามเดิม
        - ระบบ ↔ ระบบ (รวม check.lp-pao.go.th และ backend ของทุกระบบปลายทาง) ใช้ **client credentials**
          และได้ scope ตาม client scope ที่ผู้ดูแล MDM/DPO อนุมัติเป็นราย client
        - ผู้ใช้ (self-service portal, HR/DPO console) ใช้ **authorization code + PKCE** กับ Keycloak
          ซึ่ง broker ไปยัง check.lp-pao.go.th ในฐานะ OIDC Identity Provider (ทางเลือก A ในเอกสารออกแบบ)
          และได้ scope ตาม realm role (`staff`, `hr_officer`, `dpo`)
        claim ที่ MDM ใช้: `sub`, `azp` (client id), `scope`, `person_id` (user context), `roles`
      flows:
        clientCredentials:
          tokenUrl: https://iam.lp-pao.go.th/realms/lp-pao/protocol/openid-connect/token
          scopes: &scopes
            personnel:read:basic: personId, ชื่อ-สกุล, ตำแหน่ง, สังกัด, ประเภทบุคลากร, สถานะ, version
            personnel:read:contact: เบอร์โทร, อีเมล, ที่อยู่ปัจจุบัน, ผู้ติดต่อฉุกเฉิน
            personnel:read:identity: วันเกิด, เพศ, ที่อยู่ตามทะเบียนบ้าน, วันออก/หมดอายุบัตร, สถานะ ThaID
            personnel:read:employment: วันบรรจุ, ระดับ, ประวัติการดำรงตำแหน่ง, เหตุพ้นจากราชการ
            personnel:read:photo: รูปถ่ายจาก ThaID
            personnel:read:inactive: เห็น record สถานะ INACTIVE (ลาออก/โอนย้าย)
            personnel:read:pid_masked: เห็นเลขบัตรประชาชนแบบปิด เหลือ 4 หลักท้าย (employeeNoMasked)
            personnel:read:pid: ถอดรหัสเลขบัตรประชาชน (อนุมัติโดย DPO เป็นราย client)
            personnel:lookup:pid: ค้นหา personId จาก pid (ใช้ตอน migrate/เชื่อมครั้งแรก)
            personnel:self: ข้อมูลของตนเอง และแก้ไขข้อมูลติดต่อ/ความยินยอม (user context เท่านั้น)
            personnel:write:employment: HR บันทึกตำแหน่ง/สังกัด/วันบรรจุ, ระงับ/คืนสถานะ, ขอ re-verify
            personnel:provision: HR เตรียม record ล่วงหน้าและจัดการ claim request
            personnel:import: นำเข้าข้อมูลชุดจากระบบ HR เดิม
            personnel:manage:reference: >-
              เพิ่ม/แก้ master data หน่วยงานและตำแหน่ง (T10) - ต้องมี realm role `hr_master_data_admin` ควบคู่เสมอ
              (API ตรวจทั้ง scope และ role เพราะ hr-console ผูก scope นี้ให้ทุกผู้ใช้ของ client)
            personnel:manage:person: >-
              HR จัดการข้อมูลบุคคลที่ไม่ใช่ของ ThaID (PR-D2: manage-profile, expected-identity, contact, emergency-contacts, history) - ต้องมี realm role
              `hr_master_data_admin` ควบคู่เสมอ (API ตรวจทั้ง scope และ role) ไม่ผูกกับฟิลด์ใดใน response ของ Person
            sync:thaid: check.lp-pao.go.th เท่านั้น (client `check-broker`)
            events:read: อ่านฟีดเหตุการณ์
            webhook:manage: จัดการ webhook subscription ของ client ตนเอง
            audit:read: อ่าน access log และ change log (DPO / ผู้ตรวจสอบ)
            audit:review: บันทึกผลรีวิวการเปิดเลขบัตร (เฉพาะ realm role dpo - MDM API ตรวจ role เพิ่มเสมอ ไม่ใช้ scope เป็นตัวกั้นเพียงอย่างเดียว)
        authorizationCode:
...
  parameters:
...
    PersonId:
      name: personId
      in: path
      required: true
      description: internal ID (UUID) ของบุคลากร
      schema: { type: string, format: uuid }
...
    PidFormat:
      name: pidFormat
      in: query
      description: |
        `masked` = ไม่คืนเลขบัตรประชาชนเต็ม (`employeeNo`) แม้ token จะมี `personnel:read:pid` (data minimization ฝั่ง API:
        ผู้เรียกที่ต้องการแค่ `employeeNoMasked` ควรส่งค่านี้เสมอ) ไม่ส่ง = พฤติกรรมเดิมตาม scope
      schema: { type: string, enum: [masked] }
...
    Cursor:
      name: cursor
      in: query
      description: cursor จาก `page.nextCursor` ของหน้าก่อน
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: UUID ที่ผู้เรียกสร้าง ระบบจะคืนผลเดิมถ้าเรียกซ้ำภายใน 24 ชั่วโมง
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: คำขอไม่ถูกต้อง
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: scope ไม่พอ หรือ client ถูกระงับ
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: ไม่พบ (หรือไม่มีสิทธิ์เห็น record นั้น)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: ข้อมูลขัดแย้ง (เช่น version ไม่ตรง, ตำแหน่งมีผู้ครองอยู่)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unprocessable:
      description: ข้อมูลถูกรูปแบบแต่ขัดกับ business rule (เช่น หน่วยงานที่อ้างถึงไม่มี/ไม่ active)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    PersonnelType:
      type: string
      description: |
        กฎเลขที่ตำแหน่ง (`positionId`) ตามประเภทบุคลากร - MDM API บังคับทุกเส้นทางที่เขียน employment
        (POST /persons, PUT /persons/{id}/employment, POST /persons/{id}/reactivate, POST /claim-requests/{id}/resolve
        action=PROVISION, POST /sync/hr/employment-batch) ผิดกฎ → `422`:
        - **ต้องมี** (`position-required`): `CIVIL_SERVANT`, `TEACHER`, `PERMANENT_EMPLOYEE`, `TRANSFERRED_HEALTH`
        - **ห้ามมี** (`position-not-allowed`): `CONTRACT_EMPLOYEE`, `GENERAL_EMPLOYEE`, `EXPERT_EMPLOYEE`,
          `OUTSOURCE_INDIVIDUAL`, `POLITICAL_APPOINTEE`
        - **ไม่บังคับ**: `OTHER` (และประเภทใหม่ที่ยังไม่ได้กำหนดกฎ)
        batch import รายงานเป็น error รายแถว `code = POSITION_REQUIRED | POSITION_NOT_ALLOWED`
      enum:
        - CIVIL_SERVANT            # ข้าราชการองค์การบริหารส่วนจังหวัด
        - TEACHER                  # ข้าราชการครูและบุคลากรทางการศึกษา
        - PERMANENT_EMPLOYEE       # ลูกจ้างประจำ
        - CONTRACT_EMPLOYEE        # พนักงานจ้างตามภารกิจ
        - GENERAL_EMPLOYEE         # พนักงานจ้างทั่วไป
        - EXPERT_EMPLOYEE          # พนักงานจ้างผู้เชี่ยวชาญพิเศษ
        - TRANSFERRED_HEALTH       # บุคลากรถ่ายโอน (รพ.สต.)
        - OUTSOURCE_INDIVIDUAL     # จ้างเหมาบริการ (รายบุคคล ประจำในหน่วยงาน ไม่มีเลขที่ตำแหน่ง)
        - OTHER                    # อื่นๆ (ไม่เข้ากลุ่มใดใน 8 ประเภทข้างต้น ไม่มีเลขที่ตำแหน่ง)
        - POLITICAL_APPOINTEE      # ผู้ดำรงตำแหน่งทางการเมือง (คณะผู้บริหาร - ไม่มีเลขที่ตำแหน่ง)
...
    OrgUnitRef:
      type: object
      properties:
        orgUnitId: { type: string, format: uuid }
        code: { type: string }
        nameTh: { type: string }
        parentNameTh: { type: string, description: สำนัก/กอง ต้นสังกัด }
...
    Position:
      type: object
      required: [positionId, positionNo, titleTh, positionType, orgUnitId]
      properties:
        positionId: { type: string, format: uuid }
        positionNo: { type: string, description: เลขที่ตำแหน่ง }
        titleTh: { type: string }
        positionType:
          type: string
          enum: [EXECUTIVE, DIRECTOR, ACADEMIC, GENERAL, SCHOOL_DIRECTOR, SCHOOL_DEPUTY_DIRECTOR, POLITICAL]
          description: >-
            EXECUTIVE = บริหารท้องถิ่น, DIRECTOR = อำนวยการท้องถิ่น, ACADEMIC = วิชาการ, GENERAL = ทั่วไป,
            SCHOOL_DIRECTOR = ผู้อำนวยการสถานศึกษา, SCHOOL_DEPUTY_DIRECTOR = รองผู้อำนวยการสถานศึกษา,
            POLITICAL = ฝ่ายการเมือง
        orgUnitId: { type: string, format: uuid }
        isActive: { type: boolean }
...
    Employment:
      type: object
      x-source: HR
      properties:
        employmentId: { type: string, format: uuid }
        employeeNo:
          type: string
          x-required-scope: personnel:read:pid
          description: เลขประจำตัวข้าราชการ/พนักงาน = เลขบัตรประชาชนเสมอ - ดู PersonDetail.basic.employeeNo
        personnelType: { $ref: '#/components/schemas/PersonnelType' }
        position: { $ref: '#/components/schemas/Position' }
        orgUnit: { $ref: '#/components/schemas/OrgUnitRef' }
        levelCode: { type: string }
        jobTitleText:
          type: string
          maxLength: 255
          x-required-scope: personnel:read:basic
          description: ชื่อตำแหน่ง/ลักษณะงาน (ข้อความอิสระ) - ดู `EmploymentUpsert.jobTitleText`
        appointedDate: { type: string, format: date, description: วันบรรจุ }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        isCurrent: { type: boolean }
        employmentStatus:
          type: string
          enum: [ACTIVE, TRANSFERRED_OUT, RESIGNED, RETIRED, TERMINATED, DECEASED]
        separationDate: { type: [string, 'null'], format: date }
        emailWork: { type: [string, 'null'], format: email }
        hrSourceRef: { type: [string, 'null'], description: รหัสอ้างอิงในระบบ HR เดิม }
        updatedAt: { type: string, format: date-time }
...
    EmploymentChangeRequest:
      description: |
        ใช้กับ PUT /persons/{id}/employment และ POST /persons/{id}/reactivate (HR ทำด้วยมือ) = ฟิลด์ของ EmploymentUpsert + `reason` (บังคับ)
        แต่ `employeeNo` ไม่บังคับ (ไม่ส่ง = คงเลขเดิม) ส่วน POST /persons (employment ซ้อนใน PersonProvisionRequest), resolve claim และ batch import ใช้ EmploymentUpsert เดิม
      type: object
      required: [personnelType, orgUnitId, effectiveFrom, reason]
      properties:
        employeeNo:
          type: string
          maxLength: 50
          description: |
            เลขบัตรประชาชน - **ไม่ส่ง = คงค่าเดิม** ของ employment ล่าสุดของบุคคลนั้น (หน้าจอ HR ไม่มีเลขเต็มให้ส่งซ้ำ) ส่งมา = ใช้ค่าที่ส่ง
            (บุคคลที่ไม่เคยมี employment เลย -> 422 `employee-no-required`)
        personnelType: { $ref: '#/components/schemas/PersonnelType' }
        positionId:
          type: string
          format: uuid
          description: |
            เลขที่ตำแหน่ง - ไม่ระบุได้สำหรับบุคลากรที่ไม่มีตำแหน่งทางการตามโครงสร้างอัตรากำลัง
            (พนักงานจ้าง/จ้างเหมาบริการรายบุคคล)
        orgUnitId: { type: string, format: uuid }
        levelCode: { type: string, maxLength: 50 }
        jobTitleText:
          type: string
          maxLength: 255
          description: |
            ชื่อตำแหน่ง/ลักษณะงาน (ข้อความอิสระ) สำหรับบุคลากรที่ไม่มีเลขที่ตำแหน่ง - เก็บใน `employment.job_title_text` แยกจาก
            `positionId` (ไม่แทน `basic.positionTitle` และไม่อยู่ใน token claims) ระบบตัด control/bidi characters, รวมบรรทัดเป็นบรรทัดเดียว
            และ trim ก่อนบันทึก (ข้อความว่างหลังตัด = ไม่มีข้อความ) กฎตามประเภทบุคลากร (แยกจากกฎเลขที่ตำแหน่ง):
            - ประเภทที่ต้องมีตำแหน่ง (`CIVIL_SERVANT`, `TEACHER`, `PERMANENT_EMPLOYEE`, `TRANSFERRED_HEALTH`) ที่ส่งฟิลด์นี้ -> 422 `job-title-not-allowed`
            - `OTHER` ส่งทั้ง `positionId` และ `jobTitleText` -> 422 `position-and-job-title-conflict` (เลือกอย่างใดอย่างหนึ่ง หรือไม่ใส่เลย)
            - ข้อความที่มีตัวเลข 13 หลักที่เข้าข่ายเลขบัตรประชาชน (ติดกันหรือมีขีด/เว้นวรรค) -> 422 `job-title-contains-pid`
            - ยาวเกิน 255 ตัวอักษรหลัง normalize -> 422 `job-title-too-long`
            (batch import รายงานเป็น error รายแถว: `JOB_TITLE_NOT_ALLOWED`, `POSITION_AND_JOB_TITLE_CONFLICT`, `JOB_TITLE_CONTAINS_PID`, `JOB_TITLE_TOO_LONG`)
        appointedDate: { type: string, format: date }
        effectiveFrom: { type: string, format: date }
        emailWork: { type: [string, 'null'], format: email, maxLength: 255 }
        referenceDocument: { type: string, maxLength: 200, description: เลขที่คำสั่ง }
        expectedVersion:
          type: integer
          description: version ของ person ที่ผู้เรียกเห็นล่าสุด (optimistic locking) ถ้าไม่ตรงคืน 409
        reason: { type: string, minLength: 1, maxLength: 500, description: เหตุผลการเปลี่ยนแปลง (บังคับ ห้ามมีเลข 13 หลัก) }
...
    ChangeLogEntry:
      type: object
      properties:
        logId: { type: integer, format: int64 }
        changedAt: { type: string, format: date-time }
        fieldKey: { type: string, examples: [identity.reg_address_text] }
        oldValue: { description: ค่าเดิม (mask เมื่อฟิลด์เป็น RESTRICTED) }
        newValue: { description: ค่าใหม่ }
        changedBy: { type: string, enum: [THAID_SYNC, SELF, HR, HR_IMPORT, ADMIN] }
        actorSub: { type: [string, 'null'], description: "ผู้กระทำ: sub ของผู้ใช้, personId (SELF), หรือ system:thaid-sync / system:hr-import; null = แถวก่อน PR-A" }
        actorClient: { type: [string, 'null'], description: azp ของ token ที่เรียก }
        valuesHidden: { type: boolean, description: ค่าถูกปกปิดเพราะฟิลด์ชั้น CONFIDENTIAL/SENSITIVE/RESTRICTED }
        syncEventId: { type: [string, 'null'], format: uuid }
        reason: { type: [string, 'null'] }
    AuditChangeLogEntry:
      type: object
      properties:
        source: { type: string, enum: [PERSON, REFERENCE] }
        logId: { type: integer, format: int64 }
        changedAt: { type: string, format: date-time }
        personId: { type: string, format: uuid, description: เฉพาะ source=PERSON }
        recordId: { type: string, format: uuid, description: เฉพาะ source=REFERENCE }
        tableName: { type: string }
        action: { type: string, enum: [CREATE, UPDATE], description: เฉพาะ source=REFERENCE }
        fieldKey: { type: string }
        oldValue: {}
        newValue: {}
        valuesHidden: { type: boolean }
        changedBy: { type: string, enum: [THAID_SYNC, SELF, HR, HR_IMPORT, ADMIN], description: เฉพาะ source=PERSON }
        actorSub: { type: [string, 'null'] }
        actorClient: { type: [string, 'null'] }
        syncEventId: { type: [string, 'null'], format: uuid }
        reason: { type: [string, 'null'] }
...
    AccessLogEntry:
      type: object
      properties:
        accessId: { type: integer, format: int64, description: "ใช้คู่กับ accessedAt อ้างแถว (เช่น POST /audit/access-logs/{accessId}/review)" }
        accessedAt: { type: string, format: date-time }
        subjectPersonId: { type: string, format: uuid }
        actorType: { type: string, enum: [USER, SERVICE] }
        actorSub: { type: [string, 'null'] }
        clientId: { type: string }
        endpoint: { type: string }
        fieldsReturned:
          type: array
          items: { type: string }
        purposeCode: { type: [string, 'null'] }
        justification: { type: [string, 'null'] }
        requestId: { type: string }
        responseStatus: { type: integer }
        reviewStatus:
          type: [string, 'null']
          enum: [PENDING, REVIEWED, NEEDS_EXPLANATION, null]
          description: เฉพาะรายการเปิดเลขบัตร (รายการอื่น null) PENDING = ยังไม่มีผลรีวิว
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewerSub: { type: [string, 'null'] }
        reviewNote: { type: [string, 'null'] }
    PidAccessReview:
      type: object
      properties:
        reviewId: { type: integer, format: int64 }
        accessId: { type: integer, format: int64 }
        accessedAt: { type: string, format: date-time }
        status: { type: string, enum: [REVIEWED, NEEDS_EXPLANATION] }
        note: { type: [string, 'null'] }
        reviewerSub: { type: string }
        reviewedAt: { type: string, format: date-time }
    AccessAlert:
      type: object
      properties:
        alertId: { type: integer, format: int64 }
        ruleCode: { type: string, enum: [BULK_VIEW, OFF_HOURS, PID_REVEAL_FREQUENT] }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        status: { type: string, enum: [OPEN, ACK, CLOSED] }
        actorSub: { type: [string, 'null'] }
        actorClient: { type: [string, 'null'] }
        windowStart: { type: string, format: date-time }
        windowEnd: { type: string, format: date-time }
        metricCount: { type: integer }
        threshold: { type: integer }
        details: { type: object, description: ตัวเลข/พารามิเตอร์ของกฎเท่านั้น }
        detectedAt: { type: string, format: date-time }
        lastActionBy: { type: [string, 'null'] }
        lastActionAt: { type: [string, 'null'], format: date-time }
        lastActionNote: { type: [string, 'null'] }
...
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
        limit: { type: integer }
    Problem:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type: { type: string, format: uri, examples: ['https://mdm.lp-pao.go.th/problems/insufficient-scope'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: รายละเอียดรายฟิลด์สำหรับ 400
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }